Remove ADC driver header and related functions from Raspberry Pi GPIO implementation, to fix translator conflicts
