+++ coresdk/src/coresdk/raspi_gpio.cpp
@@ -4,7 +4,7 @@
 // Copyright © 2024 XQuestCode. All rights reserved.
 #include "raspi_gpio.h"
 #include "gpio_driver.h"
-#include "adc_driver.h"
+#include "raspi_adc.h"
 #include "easylogging++.h"
 #include "types.h"
 #include "backend_types.h"
@@ -17,33 +17,7 @@ namespace splashkit_lib
     int BCMpinData[] = {
         -1, -1, 2, -1, 3, -2, 4, 14, -2, 15, 17, 18, 27, -2, 22, 23, -1, 24, 10, -2, 9, 25, 11, 8, -2, 7, 0, 1, 5, -2, -6, 12, 13, -2, 19, 16, 26, 20, -2, 21};
 
-    // a function to return address based on pin number of ads7830
-    int get_ads7830_pin_address(adc_pin pin)
-    {
-        switch (pin)
-        {
-
-        case ADC_PIN_0:
-            return 0x84; // CH0
-        case ADC_PIN_1:
-            return 0x85; // CH1
-        case ADC_PIN_2:
-            return 0x86; // CH2
-        case ADC_PIN_3:
-            return 0x87; // CH3
-        case ADC_PIN_4:
-            return 0x88; // CH4
-        case ADC_PIN_5:
-            return 0x89; // CH5
-        case ADC_PIN_6:
-            return 0x8A; // CH6
-        case ADC_PIN_7:
-            return 0x8B; // CH7
-
-        default:
-            return -1; // Invalid pin
-        }
-    }
+    
 
     int boardToBCM(gpio_pin pin)
     {
@@ -279,95 +253,7 @@ namespace splashkit_lib
 #endif
     }
 
-    // ADC functions
-    adc_device open_adc(const string &name, adc_type type)
-    {
-#ifdef RASPBERRY_PI
-        if (type != ADS7830)
-        {
-            LOG(ERROR) << "Unsupported ADC type for " << name;
-            return nullptr;
-        }
-        const int default_bus = 1;
-        const int default_address = 0x48; // Default I2C address for ADS7830
-        return load_adc_device(name, default_bus, default_address, type);
-#else
-        LOG(ERROR) << "ADC not supported on this platform";
-        return nullptr;
-#endif
-    }
-    // Read the ADC value from a given channel (0-7) using a device pointer.
-    int adc_read(adc_device adc, adc_pin channel)
-    {
-#ifdef RASPBERRY_PI
-        if (adc == nullptr)
-        {
-            LOG(ERROR) << "ADC device not initialized.";
-            return -1;
-        }
-        if (channel < ADC_PIN_0 || channel > ADC_PIN_7)
-        {
-            LOG(ERROR) << "Invalid ADC channel: " << channel;
-            return -1;
-        }
-        // Convert the adc_pin enum to the corresponding channel number using get_ads7830_pin_address
-        int channel_num = get_ads7830_pin_address(channel);
-        if (channel_num == -1)
-        {
-            LOG(ERROR) << "Invalid ADC pin: " << channel;
-            return -1;
-        }
-        return read_adc_channel(adc, channel_num);
-#else
-        LOG(ERROR) << "ADC not supported on this platform";
-        return -1;
-#endif
-    }
-    
-    // Overload: read ADC value by providing the ADC device name.
-    int adc_read(const string &name, adc_pin channel)
-    {
-#ifdef RASPBERRY_PI
-        adc_device dev = adc_device_named(name);
-        if (dev == nullptr)
-        {
-            LOG(ERROR) << "ADC device " << name << " not found.";
-            return -1;
-        }
-        int channel_num = get_ads7830_pin_address(channel);
-        if (channel_num == -1)
-        {
-            LOG(ERROR) << "Invalid ADC pin: " << channel;
-            return -1;
-        }
-        return read_adc_channel(dev, channel_num);
-#else
-        LOG(ERROR) << "ADC not supported on this platform";
-        return -1;
-#endif
-    }
-    // Close an ADC device given its pointer.
-    void close_adc(adc_device adc)
-    {
-#ifdef RASPBERRY_PI
-        free_adc_device(adc);
-#else
-        LOG(ERROR) << "ADC not supported on this platform";
-#endif
-    }
-    // Overload: close an ADC device using its name.
-    void close_adc(const string &name)
-    {
-#ifdef RASPBERRY_PI
-        adc_device dev = adc_device_named(name);
-        if (dev != nullptr)
-            free_adc_device(dev);
-        else
-            LOG(WARNING) << "Attempted to close unknown ADC device: " << name;
-#else
-        LOG(ERROR) << "ADC not supported on this platform";
-#endif
-    }
+
 
     connection remote_raspi_init(const string &name, const string &host, unsigned short int port)
     {
